Fixed merge related compile errors

+++ branches/oci_dave_wchar_refactor_branch/TAO/CIAO/examples/Display/RateGen/RateGen_exec.h
@@ -61,7 +61,7 @@ namespace MyImpl
     int done_;
 
     /// The timer id we are waiting.
-    int tid_;
+    long tid_;
 
     RateGen_exec_i *pulse_callback_;
